Write the catalogue's saturated flag as int16: VOTable has no int8
The full-data build failed in write_hips_catalogue's from_table with
"dtype('int8') can not be represented in VOTable".  The new test runs
build_star_catalog itself, so the builder's own column dtypes reach
metadata.xml; it fails on the int8 version.

+++ scripts/gc_treasury_overlays.py
@@ -922,7 +922,7 @@ def build_star_catalog(M):
     columns = {
         "ra": M["ra"][ok], "dec": M["dec"][ok],
         "f212n": m, "f480m": M["m480"][ok], "color": col,
-        "saturated": M["sat"][ok].astype(np.int8),
+        "saturated": M["sat"][ok].astype(np.int16),  # VOTable has no int8
         "obs": M["who"][ok].astype(str),
         "rgb": hexes[inv.ravel()],
     }
